LayoutTests/storage/indexeddb/keypath-as-array-expected.txt

Test IndexedDB key path corner cases

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
test keypath as array
db.createObjectStore('foo', {keyPath: ['foo'], autoIncrement: true})
db.createObjectStore('foobar', {keyPath: ['foo', 'bar'], autoIncrement: true})
db.createObjectStore('empty', {keyPath: [], autoIncrement: true})
db.createObjectStore('nested', {keyPath: [[]], autoIncrement: true})
db.createObjectStore('arrayofempty', {keyPath: ['', ''], autoIncrement: true})
db.createObjectStore('arrayoffloat', {keyPath: [1.0, 2.0], autoIncrement: true})
PASS successfullyParsed is true

TEST COMPLETE
